requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
idna==3.2
multidict==5.1.0
pytz==2021.3
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
python-dotenv~=0.19.0